+++ src/info.py
@@ -2,5 +2,5 @@ class Info:
     Author = 'Andreas Pfau'
     Domain = 'apfau.de'
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.45b6"  # see https://peps.python.org/pep-0440/
-    AppDateStr = "2026-04-15"
+    AppVersionStr = "0.46b1"  # see https://peps.python.org/pep-0440/
+    AppDateStr = "2026-06-23"